+income if order -expense if buy ingredient

+++ restaurant/src/main/java/ku/cs/restaurant/dto/financial/CreateFinancialRequest.java
@@ -6,6 +6,6 @@ import java.util.Date;
 
 @Data
 public class CreateFinancialRequest {
-    private int income;
-    private int expense;
+    private double income;
+    private double expense;
 }
